gficl-app: package that abstracts the gficl window lifecycle
* src/base-app.lisp: "GFICL-APP" new package, defines a base class
GFICL-APP:BASE-APP, which must be subclassed and a bunch of methods
which must be defined on the subclass. Run the app by creating
an instance and calling START.

* examples/minimum-app.lisp: a version of examples/minimum.lisp to
demonstrate the use of the GFICL-APP package.

[250521]
* gficl.system: make base-app.lisp depend on types.lisp since it
refers to *active-objects*. (apparently depending on gficl.lisp isn't
enough, but gficl.lisp also refers to *active-objects*) todo
investigate possible defsystem bug.

[251029]
add a comment on recovering from unclean shutdown by setting
gficl::*state* to nil.

[260407]
* examples/minimum-app.lisp: (gficl-app:draw-fn): always bind the
shader.