src/vertex.lisp: optionally opt out of per vertex-data vao
putting the vertex-array object in the gficl:vertex-data object is too
estrictive: one cannot have more than one gficl:vertex-data object (say
one wants to use two: one for the position location and one for the
color location of the vertex shader)

hack it for now with a global variable gficl::*vao*.  If this is bound
gficl:vertex-data will assume that the pipeline is managing the vertex
data array and has bound it, so it will just use it and will not
deallocate it.  (note there is  code that rebinds the vertex-array to
the same value at some places. hopefully this is harmless and tmk
unknown opengl semantics won't produce nasty surprises).

* src/vertex.lisp: (vertex-data): (vao-alloced-p): new slot. (delete-gl
vertex-data): used here. (*vao*): new global variable, which the
calling pipline will manage.  (make-vertex-data-from-pointers): used
here.