src/base-app.lisp: pending-events-style: experiment with alternatives
* src/base-app.lisp: (*process-pending-events-style*): export. can be
one of :poll (default, which calls glfw:poll-events in
gficl-app:update-fn), :wait (which calls glfw:wait-events) or a number
(which is the number of seconds passed to glfw:wait-events-timout)

NOTE this requires patched cl-glfw3 which exposes
glfw:wait-events-timeout.

setting this to :wait ensures that gficl-app:update-function is called
only once (in response to new events). the :poll default means that it
is called multiple times, and it has to be written very carefully.

[251214] - replace global var *process-pending-events-style* with a
slot on gficl:base-app (introduced via mixin class idle-render-mixin).